WIP: merged getPixelBins into VoxelFunctions
