+++ Phindr3D-Python/src/Data/DataFunctions.py
@@ -167,7 +167,6 @@ class DataFunctions:
     # end mat_dot
 
     # regexpi
-    # imfinfo
     # im2col
 
     @staticmethod
