+++ Phindr3D-Python/src/VoxelGroups/VoxelFunctions.py
@@ -17,13 +17,7 @@
 from sklearn import cluster
 import numpy as np
 
-# matlab dot product
-def mat_dot(A, B, axis=0):
-    """
-    equvalent to dot product for matlab (can choose axis as well)
-    """
-    return np.sum(A.conj() * B, axis=axis)
-
+from src.Data import DataFunctions as dfunc
 
 class VoxelFunctions:
     """Static methods for finding voxel properties. Referenced from
@@ -33,21 +27,11 @@ class VoxelFunctions:
     No constructor. All parameters and methods are static.
     """
 
-    #@staticmethod
-    ## Main function for returning bin centers of pixels, supervoxels, and mega voxels
-    ## x - m x n (m is number of observations, n is number of channels/category fractions
-    ## numBins - number of categories
-    #def getPixelBins(x, numBins):
-    #    pass
-
-    def temporary(x, numBins):
-        pass
-
     @staticmethod
     # Main function for returning bin centers of pixels, supervoxels, and mega voxels
     # x - m x n (m is number of observations, n is number of channels/category fractions
     # numBins - number of categories
-    def getPixelBins(self, x, metadata, numBins):
+    def getPixelBins(x, metadata, numBins):
         # Copied from Teo's code
         Generator = metadata.Generator
         m = x.shape[0]
@@ -65,8 +49,8 @@ class VoxelFunctions:
                     randpermX)  # max_iter used to be 100. changed because bin-centers don't always match up to real values.
                 binCenters[:, :, iRandCycle] = kmeans.cluster_centers_
                 temp1 = np.add(
-                    np.array([mat_dot(binCenters[:, :, numRandRpt - 1], binCenters[:, :, numRandRpt - 1], axis=1)]).T,
-                    mat_dot(x, x, axis=1)).T  # still not sure which one of this or the next should be transposed
+                    np.array([dfunc.mat_dot(binCenters[:, :, numRandRpt - 1], binCenters[:, :, numRandRpt - 1], axis=1)]).T,
+                    dfunc.mat_dot(x, x, axis=1)).T  # still not sure which one of this or the next should be transposed
                 temp2 = 2 * (x @ binCenters[:, :, numRandRpt - 1].T)
                 a = (temp1 - temp2)
                 sumD[iRandCycle] = np.sum(np.amin(a, axis=1))
@@ -77,6 +61,7 @@ class VoxelFunctions:
                 x)  # max iter used to be 100
             binCenters = kmeans.cluster_centers_
         return np.abs(binCenters)
+    # end getPixelBins
 
 # end VoxelFunctions
